% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medecom_plots.R
\name{cve_plot3d}
\alias{cve_plot3d}
\title{Plots MeDeComSet C.V. error following kappa and lambda values tested}
\usage{
cve_plot3d(MDCset, plot.title = NULL)
}
\arguments{
\item{MDCset}{A \code{MeDeComSet} obtained with \link[MeDeCom]{runMeDeCom}.}

\item{plot.title}{A \code{character} to specify another title than the
default one for your plot.}
}
\value{
A \code{trellis} 3D lattice graph of the cross-validation error
following kappa and lambda parameters values.
}
\description{
Plots MeDeComSet C.V. error following kappa and lambda values tested
}
\examples{
# Load the example data sets
data(example.dataset, package="MeDeCom")
# Run MeDeCom (WARNING: takes 1 or 2 hours to complete)
medecom.result <- runMeDeCom(
  D, 2:10, c(0,10^(-5:-2)), NINIT=10, NFOLDS=10, ITERMAX=300, NCORES=9)
# Plot cross-validation error
cve_plot3d(MDCset = medecom.result)
}
\references{
Lutsik P. et al., MeDeCom: discovery and quantification of latent
components of heterogeneous methylomes. Genome Biol. 2017 Mar
24;18(1):55. doi: 10.1186/s13059-017-1182-6. PMID: 28340624;
PMCID: PMC5366155.

\href{https://cran.r-project.org/web/packages/lattice/lattice.pdf}{Package ‘lattice’}

\href{http://myrcodes.blogspot.com/2015/10/3d-graph-lattice-package.html}{3D Graph - Lattice Package}
}
\author{
Yoann Pageaud.
}
