% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medecom_plots.R
\name{dummy_guide}
\alias{dummy_guide}
\title{Creates a dummy legend for ggplot2 graph.}
\usage{
dummy_guide(
  labels = NULL,
  ...,
  title = NULL,
  key = draw_key_point,
  guide_args = list()
)
}
\arguments{
\item{labels}{A \code{character} vector specifying the legend keys
labels.}

\item{title}{A \code{character} specifying the legend title.}

\item{key}{A ggplot2 legend key glyph (e.g. 'draw_key_point',
'draw_key_polygon', or 'draw_key_text').}

\item{guide_args}{A \code{list} to pass more aesthetic override elements.}
}
\value{
A \code{type} object returned description.
}
\description{
Creates a dummy legend for ggplot2 graph.
}
\references{
\href{https://stackoverflow.com/questions/70977700/creating-completely-customized-legends-in-ggplot2}{Creating completely customized legends in ggplot2}
}
\author{
Yoann Pageaud.
}
\keyword{internal}
