% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medecom_plots.R
\name{LMCprop_heatmap}
\alias{LMCprop_heatmap}
\title{Plots a heatmap of LMCs proportions in all methylation samples.}
\usage{
LMCprop_heatmap(
  MDCset,
  D,
  k,
  lambda,
  dist.method = "manhattan",
  margins = ggplot2::margin(0, 1.2, 0, 0, "cm"),
  plot.title = NULL
)
}
\arguments{
\item{MDCset}{A \code{MeDeComSet} obtained with \link[MeDeCom]{runMeDeCom}.}

\item{D}{A methylation \code{matrix} you ran MeDeCom on.}

\item{k}{Am \code{integer} to specify the value of the kappa
parameter.}

\item{lambda}{A \code{numeric} to specify the value of the lambda
parameter.}

\item{dist.method}{A \code{character} to specify the method to use for
computing distances between LMCs proportions in
methylation samples(Default: dist.method = 'manhattan';
For all supported distance methods see
\link[parallelDist]{parDist}).}

\item{margins}{A ggplot2 \code{margin} object specifying widths of plot
margins
(Default: margins = ggplot2::margin(0, 1.2, 0, 0, "cm")).}

\item{plot.title}{A \code{character} to specify another title than the
default one for your plot.}
}
\value{
A \code{grob} plot of the proportion heatmap.
}
\description{
Plots a heatmap of LMCs proportions in all methylation samples.
}
\examples{
# Load the example data sets
data(example.dataset, package="MeDeCom")
# Run MeDeCom (WARNING: takes 1 or 2 hours to complete)
medecom.result <- runMeDeCom(
  D, 2:10, c(0,10^(-5:-2)), NINIT=10, NFOLDS=10, ITERMAX=300, NCORES=9)
#Plot LMC proportions heatmap
LMCprop_heatmap(MDCset = medecom.result, D = D, k = 5, lambda = 10^-2)
}
\references{
Pageaud Y. et al., BiocompR - Advanced visualizations for data
comparison.

Lutsik P. et al., MeDeCom: discovery and quantification of latent
components of heterogeneous methylomes. Genome Biol. 2017 Mar
24;18(1):55. doi: 10.1186/s13059-017-1182-6. PMID: 28340624;
PMCID: PMC5366155.
}
\author{
Yoann Pageaud.
}
