% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medecom_plots.R
\name{LMCs_dendrogram}
\alias{LMCs_dendrogram}
\title{Plots a UPGMA dendrogram clustering of LMCs with reference methylomes}
\usage{
LMCs_dendrogram(
  MDCset,
  D,
  dt.ref,
  k,
  lambda,
  cor.method = "pearson",
  plot.title = NULL,
  orientation = "v"
)
}
\arguments{
\item{MDCset}{A \code{MeDeComSet} obtained with \link[MeDeCom]{runMeDeCom}.}

\item{D}{A methylation \code{matrix} you ran MeDeCom on.}

\item{dt.ref}{A \code{data.table} containing all reference methylomes (1
methylome by column, and 1 beta value by row) you wish to
include in the clustering with the LMCs. First column must
contain methylation array probe IDs and be named
"probeIDs".}

\item{k}{Am \code{integer} to specify the value of the kappa
parameter.}

\item{lambda}{A \code{numeric} to specify the value of the lambda
parameter.}

\item{cor.method}{A \code{character} to specify the method to use for
computing correlation between LMCs and samples
(Default: cor.method = 'pearson'; For a list of all
supported correlation methods see \link[stats]{cor}).}

\item{plot.title}{A \code{character} to specify another title than the
default one for your plot.}

\item{orientation}{A \code{character} to specify whether the dendrogram
should be displayed vertically
(Default: orientation = 'v') or horizontally
(orientation = 'h').}
}
\value{
A \code{gg} plot of the dendrogram.
}
\description{
Plots a UPGMA dendrogram clustering of LMCs with reference methylomes
}
\examples{
# Load the example data sets
data(example.dataset, package="MeDeCom")
# Run MeDeCom (WARNING: takes 1 or 2 hours to complete)
medecom.result <- runMeDeCom(
  D, 2:10, c(0,10^(-5:-2)), NINIT=10, NFOLDS=10, ITERMAX=300, NCORES=9)
# Prepare the reference data.table and give rownames to the methylation
# matrix
Tref <- as.data.table(Tref)
Tref[, probeIDs := .I]
cols <- colnames(Tref)[c(6,1:5)]
Tref <- Tref[, ..cols, ]
rownames(D) <- seq(nrow(D))
# Plot a vertical clustering of LMCs and methylomes
LMCs_dendrogram(
  MDCset = medecom.result, D = D, dt.ref = data.table::as.data.table(Tref),
  k = 5, lambda = 10^-2)
# Plot an horizontal clustering of LMCs and methylomes
LMCs_dendrogram(
  MDCset = medecom.result, D = D, dt.ref = data.table::as.data.table(Tref),
  k = 5, lambda = 10^-2, orientation = 'h')
}
\references{
Lutsik P. et al., MeDeCom: discovery and quantification of latent
components of heterogeneous methylomes. Genome Biol. 2017 Mar
24;18(1):55. doi: 10.1186/s13059-017-1182-6. PMID: 28340624;
PMCID: PMC5366155.
}
\author{
Yoann Pageaud.
}
